module_3/requirements.txt
flask==3.0.0
psycopg2-binary==2.9.9
beautifulsoup4==4.12.2
lxml==4.9.3
